Moved common http client classes under http/client package
