+++ graph-commons/src/main/scala/ch/datascience/http/client/AccessToken.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package ch.datascience.clients
+package ch.datascience.http.client
 
 import ch.datascience.tinytypes.constraints.NonBlank
 import ch.datascience.tinytypes.{Sensitive, TinyType, TinyTypeFactory}
