+++ token-repository/src/main/scala/ch/datascience/tokenrepository/repository/fetching/TokenFinder.scala
@@ -21,7 +21,7 @@ package ch.datascience.tokenrepository.repository.fetching
 import cats.MonadError
 import cats.data.OptionT
 import cats.effect.{ContextShift, IO}
-import ch.datascience.clients.AccessToken
+import ch.datascience.http.client.AccessToken
 import ch.datascience.graph.events.ProjectId
 import ch.datascience.tokenrepository.repository._
