+++ token-repository/src/test/scala/ch/datascience/tokenrepository/repository/AccessTokenCryptoSpec.scala
@@ -20,7 +20,7 @@ package ch.datascience.tokenrepository.repository
 
 import java.util.Base64
 
-import ch.datascience.clients.AccessToken
+import ch.datascience.http.client.AccessToken
 import ch.datascience.crypto.AesCrypto.Secret
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.graph.events.GraphCommonsGenerators._
