+++ token-repository/src/test/scala/ch/datascience/tokenrepository/repository/fetching/FetchTokenEndpointSpec.scala
@@ -20,7 +20,7 @@ package ch.datascience.tokenrepository.repository.fetching
 
 import cats.data.OptionT
 import cats.effect.IO
-import ch.datascience.clients.AccessToken
+import ch.datascience.http.client.AccessToken
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators._
 import ch.datascience.graph.events.EventsGenerators._
