+++ webhook-service/app/ch/datascience/webhookservice/hookcreation/ProjectHookCreator.scala
@@ -19,7 +19,7 @@
 package ch.datascience.webhookservice.hookcreation
 
 import cats.effect.IO
-import ch.datascience.clients.{AccessToken, IORestClient}
+import ch.datascience.http.client.{AccessToken, IORestClient}
 import ch.datascience.graph.events.ProjectId
 import ch.datascience.webhookservice.IOContextShift
 import ch.datascience.webhookservice.config.IOGitLabConfigProvider
