+++ webhook-service/app/ch/datascience/webhookservice/hookcreation/UserInfoFinder.scala
@@ -19,7 +19,7 @@
 package ch.datascience.webhookservice.hookcreation
 
 import cats.effect.IO
-import ch.datascience.clients.{AccessToken, IORestClient}
+import ch.datascience.http.client.{AccessToken, IORestClient}
 import ch.datascience.graph.events._
 import ch.datascience.webhookservice.IOContextShift
 import ch.datascience.webhookservice.config.IOGitLabConfigProvider
